requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
certifi==2024.7.4
distro==1.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
openai==1.40.3
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
wheel==0.44.0
fastapi~=0.112.0
uvicorn~=0.30.5
starlette~=0.37.2